+++ packages/gestalt-codemods/1.15.0-1.16.0/__testfixtures__/deprecatedPadding-box-replace.input.js
@@ -1,0 +1,24 @@
+// @flow
+import React from 'react';
+import { Box, Box as Boxie } from 'gestalt';
+
+export default function TestBox() {
+  return (
+    <Boxie deprecatedPadding={2}>
+      <Box
+        alignItems="center"
+        deprecatedPadding={{ x: 2 }}>
+          <Box />
+      </Box>
+      <Box
+        deprecatedPadding={{ y: 2 }}>
+          <Box />
+      </Box>
+      <Box
+        alignItems="center"
+        deprecatedPadding={{ x: 2, y: 3 }}>
+          <Box />
+      </Box>
+    </Boxie>
+  );
+}