+++ packages/gestalt-codemods/1.15.0-1.16.0/__tests__/deprecatedMargin-box-replace.test.js
@@ -1,0 +1,21 @@
+import { defineTest } from 'jscodeshift/dist/testUtils.js';
+
+jest.mock('../deprecatedMargin-box-replace', () => {
+  return Object.assign(
+    require.requireActual('../deprecatedMargin-box-replace'),
+    {
+      parser: 'flow',
+    }
+  );
+});
+
+describe('deprecatedMargin-box-replace', () => {
+  ['deprecatedMargin-box-replace'].forEach(test => {
+    defineTest(
+      __dirname,
+      'deprecatedMargin-box-replace',
+      { quote: 'single' },
+      test
+    );
+  });
+});